mission17/requirements.txt
streamlit==1.29.0
streamlit-drawable-canvas==0.9.3
numpy==1.26.4
Pillow==10.0.1
onnxruntime==1.19.2
requests==2.31.0